Remove post endpoint and component
Post request did not worked properly.
When filter was applied not the selected item was added to the cart.
Most probably the issue was related with the hidden tag and form submission.
Comonent was removed because the Index view is simple enough. No need to break down the logic.
Add to cart feature is solved with get request.